Rename ActivityFeed view to ConfSelector view
