+++ src/main/resources/com/gluonhq/otn/otn.properties
@@ -56,7 +56,7 @@ OTN.VIEW.SPEAKER=Speaker
 OTN.VIEW.SPEAKERS=Speakers
 OTN.VIEW.SESSION=Session
 OTN.VIEW.SESSIONS=Content Catalog
-OTN.VIEW.ACTIVITYFEED=Conference Selector
+OTN.VIEW.CONFSELECTOR=Conference Selector
 OTN.VIEW.SPLASH=Splash
 OTN.VIEW.EULA=EULA
 OTN.VIEW.PROFILE=Profile
