--- src/main/resources/com/gluonhq/otn/views/activityfeed.fxml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import com.gluonhq.charm.glisten.mvc.View?>
-<?import com.gluonhq.charm.glisten.control.CardPane2?>
-
-<View fx:id="activityFeedView" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.gluonhq.otn.views.ActivityFeedPresenter">
-</View>
